% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{add_list_total}
\alias{add_list_total}
\title{Add a TOTAL to a list of counts.}
\usage{
add_list_total(l)
}
\arguments{
\item{l}{The existing list.}
}
\value{
A new list with a TOTAL at the end.
}
\description{
Add a TOTAL to a list of counts.
}
