% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_is_grant_esi_related}
\alias{annotate_is_grant_esi_related}
\title{Annotate if grant is ESI-related (compute from grant table).}
\usage{
annotate_is_grant_esi_related(grants)
}
\arguments{
\item{grants}{The grants table}
}
\value{
A grants table with a new logical column, is_grant_esi_related.
}
\description{
An Early Stage investigator (ESI) is noted in the Partnership Role for an
individual associated with a grant. Note that a person's Partnership Role can
change over the course of different grant submissions. See is_esi_investigator()
on how this works.
}
\details{
If an ESI is associated with a grant (in any grant role), then the grant
is considered an ESI-related grant.
}
