% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_is_grant_phsu_cancer_related}
\alias{annotate_is_grant_phsu_cancer_related}
\title{Title}
\usage{
annotate_is_grant_phsu_cancer_related(grants)
}
\arguments{
\item{grants}{}
}
\value{

}
\description{
Title
}
