% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_pg_year}
\alias{annotate_pg_year}
\title{Assign the Partnership Grant year to grant submitted, grant funded.}
\usage{
annotate_pg_year(d)
}
\arguments{
\item{d}{The tibble to operate on.}
}
\value{
A modified tibble with date submitted/funded as U54 years.
}
\description{
Assign the Partnership Grant year to grant submitted, grant funded.
}
