% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{apply_u54reportr_flextable_style}
\alias{apply_u54reportr_flextable_style}
\title{Apply default style}
\usage{
apply_u54reportr_flextable_style(table)
}
\arguments{
\item{table}{}
}
\description{
This is a temporary hack until I figure out flextable themes.
}
