% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_investigators_by_role}
\alias{arrange_investigators_by_role}
\title{Arrange investigator table by role}
\usage{
arrange_investigators_by_role(investigators)
}
\arguments{
\item{investigators}{An investigators tibble}
}
\value{
A sorted investigators table, ordered by investigator role.
}
\description{
Grant roles have a certain hierarchy (PI, Co-I, postdoc, etc). This function
will sort the investigators by this predefined hierarchy.
}
\details{
The package variable \code{role_order} defines the order of roles to be considered.
}
