% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{assign_cores}
\alias{assign_cores}
\title{Convert core assignments to merged field.}
\usage{
assign_cores(g)
}
\arguments{
\item{g}{Grant table}
}
\value{
A modified tibble with U54 Core Support variable.
}
\description{
Convert core assignments to merged field.
}
