% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{assign_grant_tags}
\alias{assign_grant_tags}
\title{Convert grant tags to printable text.}
\usage{
assign_grant_tags(d)
}
\arguments{
\item{d}{Tibble of grants}
}
\value{
A modified tibble with Grant Tags and Grant Status fields.
}
\description{
Convert grant tags to printable text.
}
