% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{clean_redcap_data}
\alias{clean_redcap_data}
\title{Clean the imported redcap file for later use.}
\usage{
clean_redcap_data(d)
}
\arguments{
\item{d}{RedCAP data to clean.}
}
\value{
A cleaned version of the RedCAP data.
}
\description{
The default is to create factors for variables and to
have labels that represent the questions. This library
is for reporting of the data in the database rather than
keeping the specific question text. Therefore, the
labels are stripped and the factors are converted to text.
}
