% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_collapse.R, R/format_investigators.R
\name{collapse_investigators}
\alias{collapse_investigators}
\title{Collapse investigator list into string (deprecated)}
\usage{
collapse_investigators(grants)

collapse_investigators(grants)
}
\arguments{
\item{grants}{the grants tibble}

\item{investigators}{The investigators to collapse to a string.}

\item{label}{A field name of the tibble or a fixed string for the list name.}

\item{sep}{The string to use in separating individual investigators.}
}
\value{


A grants tibble with a new column (Investigators).
}
\description{
For a given tibble of investigators, assuming it represents a single group,
create a named list consisting of a single string representing all of the investigators
(separated by sep). The name of the list is the either a variable in the data or a
text string.

There are (potentially) multiple investigators per grant. This function will
reduce this to a single text string per grant, adding it to the grants table as the
\code{Investigators} field.
}
\details{
The primary purpose of this function is to collapse a tibble of investigators on the
summary_field (NB: should this actually just be a function?) into a single string,
separated by \code{sep}. The string value can also have a name associated with it, given
by \code{label}. If \code{label} is a variable in the investigator tibble and all values of that
variable are the same, then that value will be used. Some example is needed.

For instance, if \code{label} is \code{Institution} and 1) this is a variable in the tibble and
2) all rows have a single value for this variable (for instance, \code{MCC}), then the list
will be labeled as \code{MCC}=string.

The label field allows you to specify one of the tibble columns to use as the label, or
if it's not found, then it is assumed to be a string to use for this purpose.

Given the grants table, the current implementation is to have investigators (a tibble)
nested inside of the grant tibble. This is convenient to retain the dplyr semantics,
but at some point the investigators have to be collapsed to a single string. This is that
function.

The job here is to take the investigators out, format the investigators per the style
guidelines, then store the resulting strings back into the grants table.
}
