% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_investigators.R
\name{combine_investigator_groups_as_flextable}
\alias{combine_investigator_groups_as_flextable}
\title{Format a list of investigator flextable chunks.}
\usage{
combine_investigator_groups_as_flextable(l, sep = "\\n")
}
\arguments{
\item{l}{A list of investigator flextable chunks}
}
\value{
A flextable paragraph representing all of the chunks.
}
\description{
Given a list of possible grouped investigators as flextable chunks, format this into a
single flextable paragraph.
}
