% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_investigators.R
\name{combine_investigator_groups_as_text}
\alias{combine_investigator_groups_as_text}
\title{Format a list of investigator text chunks.}
\usage{
combine_investigator_groups_as_text(l, sep = " | ", ...)
}
\arguments{
\item{l}{A list of investigator text chunks}
}
\value{
A string representing all investigators
}
\description{
Given a list of possibly grouped investigators as text chunks, format this into a single
string as text.
}
\details{
Combining investigators into a single string is a two-step process. First, each group
can be combined into a text chunk. This list of text chunks would then be further combined
into a single string (this function). The idea is that this process can be done for
both flextables and text using the same type of functions. This function handles
text output.
}
