% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_date}
\alias{convert_date}
\title{Convert date from redcap (m/d/y format).}
\usage{
convert_date(s)
}
\arguments{
\item{s}{A character vector to convert.}
}
\value{
A Date object.
}
\description{
Convert date from redcap (m/d/y format).
}
