% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partnership_grant.R
\name{convert_date_to_pg_year}
\alias{convert_date_to_pg_year}
\title{Convert a calendar date to a Partnership Grant year.}
\usage{
convert_date_to_pg_year(d)
}
\arguments{
\item{d}{A lubridate-style date.}
}
\value{
A string representing the pg year, or NA if not defined.
}
\description{
Grant years are often not aligned to calendar years or to fiscal years.
They represent their own internal clock. Here, the Partnership Grant
is the reason for reporting on activity within the Partnership. Therefore,
it is important to track the Partnership Grant (pg) year.
}
\details{
The logic implemented in this function is straight-forward. If the date is
less than the start date of a given pg year, then it is the previous
year. This works since each year is tested in increasing order.

Note that dates starting earlier than Y01 start date are considered
Y00.
}
