% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{convert_redcap_data}
\alias{convert_redcap_data}
\title{Convert the RedCAP data into the grant tibble.}
\usage{
convert_redcap_data(d)
}
\arguments{
\item{d}{The RedCAP data to convert.}
}
\value{
A grant tibble.
}
\description{
Convert the RedCAP data into the grant tibble.
}
