% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Count number of research products based on different criteria}
\description{
Counting totals can be a bit difficult in complex
data because the logic of what should be counted may not always
be obvious. The \code{count_} functions in this package are designed
to encode this logic and memorialize it for future reports.
\itemize{
\item \code{\link[=filter_between]{filter_between()}}
\item \code{\link[=filter_in_year]{filter_in_year()}}
\item \code{\link[=filter_supported]{filter_supported()}}
\item \code{\link[=filter_core_supported]{filter_core_supported()}}
\item \code{\link[=filter_by_tag]{filter_by_tag()}}
}
}
\note{
Ideally the \code{count} routines will only count existing records and
not try to do additional filtering, but this may be necessary (or desirable). If
filtering, then it should use the \code{\link[=filter]{filter()}} section of this package.
}
