% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_current_esi_mentions_in_grant}
\alias{count_current_esi_mentions_in_grant}
\title{Title}
\usage{
count_current_esi_mentions_in_grant(g)
}
\arguments{
\item{g}{}
}
\description{
Title
}
\details{
This is a function to expand out the investigators to find ESI's,
then provide a total count of these. Note this will mean multiple
entries per grant.
}
