% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_esi_creators}
\alias{count_esi_creators}
\title{Count the ESI's in a creator table}
\usage{
count_esi_creators(.x)
}
\arguments{
\item{.x}{An investigator or author table}
}
\value{
The count of the number of ESI's
}
\description{
In either a grant investigators or publication authors table, count
the number of ESIs.
}
\details{
A grant has a table of investigators and a publication has a table
of authors. In either case, each creator can have a Partnership Role of
ESI. This function will count the number of such ESI's for the grant/publication.
}
\note{
Empty tables can be a little tricky, so there is special code to return
a count of 0 if the table is empty.
}
