% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_grants_funded}
\alias{count_grants_funded}
\title{Counts grants funded}
\usage{
count_grants_funded(d)
}
\arguments{
\item{d}{Grants table}
}
\value{
Total grants funded
}
\description{
Counts grants funded
}
