% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_grants_funded_by_year}
\alias{count_grants_funded_by_year}
\title{Count funded submissions
This has some subtlety. We need to actually consider the year it is funded, as opposed
to the year it was submitted.}
\usage{
count_grants_funded_by_year(d)
}
\arguments{
\item{yr}{}
}
\value{

}
\description{
Count funded submissions
This has some subtlety. We need to actually consider the year it is funded, as opposed
to the year it was submitted.
}
