% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_submissions}
\alias{count_submissions}
\title{Count number of grant submissions}
\usage{
count_submissions(d)
}
\arguments{
\item{d}{Grants table}
}
\value{
The count of number of grant submissions
}
\description{
Count number of grant submissions
}
