% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_submissions_by_year}
\alias{count_submissions_by_year}
\title{Count grant submissions by year}
\usage{
count_submissions_by_year(d)
}
\arguments{
\item{d}{Grants table}
}
\value{
A named vector of \code{year}=\code{count}.
}
\description{
Summarize grant submissions (total) by U54 fiscal year.
}
