% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{create_info_card}
\alias{create_info_card}
\title{Create info card}
\usage{
create_info_card(.x, tag)
}
\arguments{
\item{.x}{A table of information to extract a row with \code{tag}=tag.}

\item{tag}{The tag for the \verb{tag column,}}
}
\description{
Create a info card, which is a plot taking the whole screen, based on
the input table \code{.x} and field \code{tag}.
}
\details{
This function is a little brittle. It assumes there is a tag
column (to filter on) and a Description, Count and Color fields for
the plot.
}
