% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{create_investigators_table}
\alias{create_investigators_table}
\title{Create a table specific for investigator information.}
\usage{
create_investigators_table(data)
}
\arguments{
\item{data}{Raw source redcap data}
}
\value{
A tibble corresponding to investigators.
}
\description{
Investigator data is embedded in the raw data as extra rows with
no grant information, but investigator information. We extract out
the investigator information and do a bit of data wrangling to clean
it up.
}
\details{
The linkage between investigators and grants is the record_id (renamed to grant_id).
}
