% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{datestamp}
\alias{datestamp}
\title{Return today's date as YYMMDD for printing.}
\usage{
datestamp(tzone = "")
}
\value{
A string representing today's date.
}
\description{
Return today's date as YYMMDD for printing.
}
