% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_funding_dates}
\alias{derive_funding_dates}
\title{Add funding dates}
\usage{
derive_funding_dates(d)
}
\arguments{
\item{d}{Tibble representing the grant table.}
}
\value{
A modified tibble with Funding Start Date and Funding End Date.
}
\description{
Add funding dates
}
