% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_funding_source}
\alias{derive_funding_source}
\title{Create a composite string for funding source (agency/program).}
\usage{
derive_funding_source(grant_agency, funding_program)
}
\arguments{
\item{funding_program}{}
}
\value{

}
\description{
This function formats the grant agency and funding program into
a combined string, with an optional slash if possible.
}
