% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_investigator_summary}
\alias{derive_investigator_summary}
\title{Derive investigator name and roles as single text.}
\usage{
derive_investigator_summary(investigators, ...)
}
\arguments{
\item{investigators}{The investigators tibble for a grant.}
}
\value{
A formatted string (or vector of strings) representing investigators
}
\description{
An investigator can have a partnership role and a grant role. This
formats the information in many different ways.
}
\details{
An investigator is a person associated with a grant. There can be one or
more investigators on each grant, and they may have different roles on
each grant. Additionally, the investigator has a partnership role which
may change over time. Thus, each grant has a tibble associated with the
investigators for that grant and include descriptions of their name,
grant role and partnership role.

This function produces a text summary of this information so that it could
be printed out succinctly. There are many different parameters that can be
tweaked (see \code{format_investigator_name} for many formatting options passed on
through this function).
}
\seealso{
\code{\link[=format_investigator_name]{format_investigator_name()}} for flags to format the investigator name.
}
