% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Dictionary routines}
\description{
There is a data dictionary (metadata) that can be extracted from REDCap,
documenting the instruments and variables. This is important for mapping
some fields that are either in coded, checkbox or other forms.
}
\details{
The functions in \code{dictionary.R} provide ways to extract information from
the data dictionary and to translate survey data (using the dictionary)
into usable form.
}
