% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.convert_checkbox_to_flags}
\alias{.convert_checkbox_to_flags}
\title{Convert checkbox fields to indicator variables}
\usage{
.convert_checkbox_to_flags(.x, raw_prefix, indicator_prefix, dictionary)
}
\arguments{
\item{.x}{Data frame to operate on}

\item{raw_prefix}{The checkbox variable prefix (excluding the ___) in the original table}

\item{indicator_prefix}{The prefix to use in for human-readable flags. It will be prepended
to the variable names, with a '_' separating it.}
}
\value{
Data frame with new columns representing human-readable indicators.
}
\description{
A checkbox field consists of many different fields with values Checked
and Unchecked. This function converts this to more human-readable form.
}
\details{
This function performs several related tasks to transform a
checkbox (set of columns) into several human-readable columns. It is
highly customized for the specific data input and not a general-purpose
function.
}
