% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_publications.R
\name{.convert_publication_date}
\alias{.convert_publication_date}
\title{Convert publication dates to date format}
\usage{
.convert_publication_date(s)
}
\arguments{
\item{s}{A vector of strings representing dates (hopefully)}
}
\value{
A vector of dates
}
\description{
Convert publication dates to date format
}
\details{
Conversion of hand-entered publication dates can be very tricky,
as there are many different formats possible. This functionality was
extracted to its own function so that this can be tweaked as necessary.
}
