% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_publications.R
\name{.create_authors_table}
\alias{.create_authors_table}
\title{Create a table specific for author information}
\usage{
.create_authors_table(pl)
}
\arguments{
\item{.x}{Raw source redcap data}
}
\value{
A tibble corresponding to authors.
}
\description{
Author data is embedded in the raw data as extra rows with
no publication information, but author information. We extract out
the author information and do a bit of data wrangling to clean
it up.
}
\details{
The linkage between authors and publications is the pub_id.
}
