% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.create_dictionary}
\alias{.create_dictionary}
\title{Create dictionary from download object}
\usage{
.create_dictionary(.x)
}
\arguments{
\item{.x}{Download object (from \code{\link[REDCapR:redcap_metadata_read]{REDCapR::redcap_metadata_read()}}).}
}
\description{
Create dictionary from download object
}
