% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.derive_creator_summary}
\alias{.derive_creator_summary}
\title{Derive creator name and roles as single text}
\usage{
.derive_creator_summary(name, partnership_role, role = NA, ...)
}
\arguments{
\item{investigators}{The investigators tibble for a grant.}
}
\value{
A formatted string (or vector of strings) representing investigators
}
\description{
A creator (investigator or author) can have a partnership role and a
specific role. This function formats the information for output.
}
\details{
A creator is a person associated with a work (grant or publication). There
can be one or
more creators, and they may have different roles on
each work. Additionally, the creator has a partnership role which
may change over time. Thus, each work has a tibble associated with the
investigators for that work and include descriptions of their name,
grant role and partnership role.

This function produces a text summary of this information so that it could
be printed out succinctly. There are many different parameters that can be
tweaked (see \code{format_investigator_name} for many formatting options passed on
through this function).
}
\seealso{
\code{\link[=format_investigator_name]{format_investigator_name()}} for flags to format the investigator name.
}
