% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_grants.R
\name{.derive_fiscal_year_summary}
\alias{.derive_fiscal_year_summary}
\title{Format the fiscal year as submitted (funded).}
\usage{
.derive_fiscal_year_summary(submitted, funded)
}
\arguments{
\item{submitted}{The year submitted}

\item{funded}{The year funded (or NA if not funded)}
}
\value{
A string combining the two
}
\description{
Format the fiscal year as submitted (funded).
}
\details{
Partnership grant years are numbered like Y05, Y06. These are important when
tracking outcomes in the grant. However, it can be a little confusing that there
is the date a grant was submitted vs the date a grant was funded. These can often
be years apart. This function takes both the year submitted (everyone has this) and
the year funded (not everyone has this) and formats a combination of the two.

The format of the output is

\if{html}{\out{<div class="sourceCode">}}\preformatted{YrSubmitted (YrFunded)
}\if{html}{\out{</div>}}
}
\examples{

derive_fiscal_year_summary("Y05","Y06")

}
