% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.is_esi_related}
\alias{.is_esi_related}
\title{Determine if any authors are ESI's}
\usage{
.is_esi_related(.x)
}
\arguments{
\item{.x}{}
}
\description{
This function provides a way, in mutate, to test if any
of the authors are ESI's.
}
\details{
It is a little tricky, since the authors are nested, to test
if any of the authors are ESI's. This code is private to the library
since the whole purpose of the function is to create an indicator
variable so that we don't need to do this regularly.
}
