% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_grants.R
\name{.is_grant_joint}
\alias{.is_grant_joint}
\title{Annotate if grant is joint (compute from grant table)}
\usage{
.is_grant_joint(.x)
}
\arguments{
\item{grants}{The grants table}
}
\value{
A grants table with a new logical column, is_grant_joint.
}
\description{
A grant is a joint grant if investigators from the two
primary institutions (PHSU and MCC) have roles on the
grant. This function determines if a grant is a joint
grant, taking as input a grant table and adding a new
column to the table called "is_grant_joint" which is a
logical.
}
