% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.pg_year}
\alias{.pg_year}
\title{Annotate date(s) to program grant year}
\usage{
.pg_year(.x)
}
\arguments{
\item{.x}{A date or vector of dates to map.}
}
\value{
A vector of characters representing program grant years.
}
\description{
This function operates on a date or list of dates, returning
the corresponding program grant year (e.g., Y01).
}
