% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_publications.R
\name{.remove_authors_from_redcap_table}
\alias{.remove_authors_from_redcap_table}
\title{Remove publication fields from publication tibble}
\usage{
.remove_authors_from_redcap_table(.x)
}
\arguments{
\item{d}{A tibble containing publications and authors.}
}
\value{
A modified tibble with publication entries removed.
}
\description{
The publication table includes rows (and columns)
representing authors (one row per author). This function keeps
the rows and columns belonging to authors specifically.
}
