% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_grants.R
\name{.remove_grants}
\alias{.remove_grants}
\title{Remove grant entries from a tibble.}
\usage{
.remove_grants(.x)
}
\arguments{
\item{.x}{A tibble containing investigators and grants.}
}
\value{
A modified tibble with grant entries removed.
}
\description{
Remove grant entries from a tibble.
}
