% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_grants.R
\name{.remove_investigators}
\alias{.remove_investigators}
\title{Remove investigator entries from a tibble.}
\usage{
.remove_investigators(.x)
}
\arguments{
\item{.x}{A tibble containing investigators and grants.}
}
\value{
A modified tibble with investigator entries removed.
}
\description{
Remove investigator entries from a tibble.
}
