% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filtering grants and publications}
\description{
Filtering grants and publications
}
\details{
This contains code for filtering the grant table based on
various criteria. It is easier
to implement these as functions, rather than remember the specific variables to filter on.
The logic can be implemented consistently this way.
}
