% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_core_supported}
\alias{filter_core_supported}
\title{Filter grant table on those grants with any core support.}
\usage{
filter_core_supported(d)
}
\arguments{
\item{d}{The grant table}
}
\value{
A filtered grant table with only grants having grant support.
}
\description{
Filter grant table on those grants with any core support.
}
