% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_grants_current_esi_related}
\alias{filter_grants_current_esi_related}
\alias{filter_grants_former_esi_related}
\title{Filter grant table to grants that are ESI-related.}
\usage{
filter_grants_current_esi_related(d)

filter_grants_former_esi_related(d)
}
\arguments{
\item{d}{A grant table}
}
\value{
A filtered grant table with only ESI-related grants.
}
\description{
Filter grant table to grants that are ESI-related.
}
\section{Functions}{
\itemize{
\item \code{filter_grants_former_esi_related()}: Filter publications in Partnership Grant year yr

}}
