% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_grants_in_preparation}
\alias{filter_grants_in_preparation}
\title{Filter the grants table to those in preparation.}
\usage{
filter_grants_in_preparation(d)
}
\arguments{
\item{d}{The grants table from u54reportr}
}
\value{
A filtered grants table, limited to those grants in preparation.
}
\description{
The grant table includes all grants (unless previously filtered). Some
grants are in preparation, under review, funded, etc. This is a filter
to limit to those in preparation.
}
