% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_grants_joint}
\alias{filter_grants_joint}
\title{Filter grants table to only grants with investigators from both MCC and PHSU.}
\usage{
filter_grants_joint(grants)
}
\arguments{
\item{grants}{A grants table}
}
\value{
A filtered grants table that consists of joint grants
}
\description{
Filter grants table to only grants with investigators from both MCC and PHSU.
}
