% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_grants_not_funded}
\alias{filter_grants_not_funded}
\title{Filter grant list to grants not funded.}
\usage{
filter_grants_not_funded(d)
}
\arguments{
\item{d}{The grant table.}
}
\value{
A table of grants not funded.
}
\description{
Filter grant list to grants not funded.
}
