% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_grants_pending_review}
\alias{filter_grants_pending_review}
\title{Filter the grants table to those Pending Review.}
\usage{
filter_grants_pending_review(d)
}
\arguments{
\item{d}{The grants table from u54reportr}
}
\value{
A filtered grants table, limited to those grants in pending review.
}
\description{
The grant table includes all grants (unless previously filtered). Some
grants are in preparation, under review, funded, etc. This is a filter
to limit to those pending review.
}
