% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_grants_phsu_cancer_related}
\alias{filter_grants_phsu_cancer_related}
\title{Filter grants table to PHSU Cancer-related grants.}
\usage{
filter_grants_phsu_cancer_related(grants)
}
\arguments{
\item{grants}{A grants table}
}
\value{
A filtered grants table that is PHSU Cancer Related
}
\description{
This filter will reduce the grant table to those grants with a Grant Tags
field including PHSU Cancer-related.
}
