% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_grants_rtype}
\alias{filter_grants_rtype}
\title{Filter grants table to only grants that are R-type.}
\usage{
filter_grants_rtype(grants)
}
\arguments{
\item{grants}{A grants table}
}
\value{
A filtered grants table that consists of R-type grants
}
\description{
Filter grants table to only grants that are R-type.
}
