% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_grants_submitted_between}
\alias{filter_grants_submitted_between}
\title{Filter to grants submitted between start and end.}
\usage{
filter_grants_submitted_between(
  d,
  start,
  end = format(Sys.time(), "\%Y\%m\%d")
)
}
\arguments{
\item{d}{The grants table}

\item{start}{Start date (YYYY-MM-DD)}

\item{end}{End date (YYYY-MM-DD)}
}
\value{
A filtered grants table.
}
\description{
This function filters the grants table to only grants that are submitted between the
specified start and end date. The dates must be input as YYYY-MM-DD format.
}
\details{
Note that
if you do not specify an end date then today's date is used.

Note that the dates are inclusive (meaning they include the date specified).
}
