% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_grants_submitted_in_year}
\alias{filter_grants_submitted_in_year}
\title{Filter grant table d to submissions in Partnership Grant year yr.}
\usage{
filter_grants_submitted_in_year(d, yr)
}
\arguments{
\item{d}{Grant table}

\item{yr}{Partnership Grant Year to filter by.}
}
\value{
A grant table filtered to submissions from year yr.
}
\description{
The grant table contains grants from all time periods. However, reporting
often occurs by specific Grant Year. The Grant Year is different from the
calendar year and fiscal year, and is specifically defined for the grant.
}
\details{
Additionally, a grant can be submitted (more than once) prior to being funded
(if it ever does get funded). This function handles submissions, rather than
funding.

Thus, this function filters a grant table to those that were submitted
in Partnership Grant year yr.

NOTE: This function is vector-aware, meaning that multiple grant years can be
included.
}
