% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R, R/filter_grants.R
\name{filter_in_year}
\alias{filter_in_year}
\alias{filter_grants_submitted_in_year}
\alias{filter_grants_funded_in_year}
\title{Filter records in Partnership Grant year yr.}
\usage{
filter_in_year(.x, var, yr = pg_project_years())

filter_grants_submitted_in_year(.x, ...)

filter_grants_funded_in_year(.x, ...)
}
\arguments{
\item{var}{Variable to use for grant year}

\item{yr}{Partnership Grant Year to filter by.}

\item{d}{tibble of records}
}
\value{
A tibble filtered to submissions from year yr.
}
\description{
Records are from all time periods. However, reporting
often occurs by specific Grant Year. The Grant Year is different from the
calendar year and fiscal year, and is specifically defined for the
partnership grant.
}
\details{
Thus, this function filters records to those that were submitted
in Partnership Grant year yr.

NOTE: This function is vector-aware, meaning that multiple grant years can be
included.
}
\section{Functions}{
\itemize{
\item \code{filter_grants_submitted_in_year()}: Filter grants submitted in year

\item \code{filter_grants_funded_in_year()}: Filter grants with funding starting in PG year yr.

}}
