% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u54.R
\name{filter_moffitt_health_disparities}
\alias{filter_moffitt_health_disparities}
\alias{filter_pub_moffitt_health_disparities}
\alias{filter_grants_moffitt_health_disparities}
\title{Filter grants based on Moffitt Health Disparities}
\usage{
filter_moffitt_health_disparities(.x)

filter_pub_moffitt_health_disparities(.x)

filter_grants_moffitt_health_disparities(.x)
}
\arguments{
\item{.x}{}
}
\description{
Filter grants based on Moffitt Health Disparities
}
\section{Functions}{
\itemize{
\item \code{filter_pub_moffitt_health_disparities}: Filter grants on Moffitt Health Disparities

\item \code{filter_grants_moffitt_health_disparities}: Filter grants on Moffitt Health Disparities
}}

