% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_pg_years_funded}
\alias{filter_pg_years_funded}
\title{Filter grant table d to grants that start funding within the U54 fiscal years grant_years.}
\usage{
filter_pg_years_funded(d, grant_years)
}
\arguments{
\item{d}{The grant table}

\item{grant_years}{A list of PG fiscal years (e.g., Y01, Y02) to filter to.}
}
\value{
A filtered grant table with only grants whose funding started in grant_years.
}
\description{
Starting from the grant table, filter the table to only include
grants that started funding in specific Program Grant fiscal years (listed as
Y01, Y02, ...). Note: grant funding can occur in a different
fiscal year from grant submission.
}
