% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u54.R
\name{filter_phsu_cancer_related}
\alias{filter_phsu_cancer_related}
\alias{filter_pub_phsu_cancer_related}
\alias{filter_grants_phsu_cancer_related}
\title{Filter table to PHSU Cancer-related projects}
\usage{
filter_phsu_cancer_related(.x)

filter_pub_phsu_cancer_related(.x)

filter_grants_phsu_cancer_related(.x)
}
\arguments{
\item{.x}{A  table}
}
\value{
A filtered grants table that is PHSU Cancer Related
}
\description{
This filter will reduce the table to those entries that are PHSU Cancer related
}
\section{Functions}{
\itemize{
\item \code{filter_pub_phsu_cancer_related}: Filter pubs on  PHSU Cancer-related

\item \code{filter_grants_phsu_cancer_related}: Filter grants on PHSU Cancer-related
}}

