% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_publications.R
\name{filter_publications_former_esi_related}
\alias{filter_publications_former_esi_related}
\alias{filter_pubs_former_esi_related}
\title{Filter publications to 'former ESI' related status}
\usage{
filter_publications_former_esi_related(.x)

filter_pubs_former_esi_related(.x)
}
\description{
Filter publications to 'former ESI' related status
}
\seealso{
\itemize{
\item \code{\link[=filter_publications]{filter_publications()}} for publications-based filtering.
\item \code{\link[=is_publication_former_esi_related]{is_publication_former_esi_related()}} for ESI-related status.
\item \code{\link[=is_publication_current_esi_related]{is_publication_current_esi_related()}} for current ESI status.
}
}
