% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_u54_fiscal_years_submitted}
\alias{filter_u54_fiscal_years_submitted}
\title{Filter grant table d to grants submitted within the Program Grant fiscal years grant_years.}
\usage{
filter_u54_fiscal_years_submitted(d, grant_years)
}
\arguments{
\item{d}{The grant table}

\item{grant_years}{A list of PG fiscal years (e.g., Y01, Y02) to filter to.}
}
\value{
A filtered grant table with only grants that were submitted in grant_years.
}
\description{
Starting from the grant table, filter the table to only include
grants that were submitted in specific U54 fiscal years (listed as
Y01, Y02, ...). Note: grant funding can occur in a different
fiscal year from grant submission.
}
