% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pubs.R
\name{format_author_as_chunks}
\alias{format_author_as_chunks}
\title{Format an author as a list of chunks for flextable}
\usage{
format_author_as_chunks(...)
}
\arguments{
\item{x}{}
}
\description{
Format an author (defined in a list) as a list of
formatted chunks suitable for inserting into a \code{\link[flextable:as_paragraph]{flextable::as_paragraph()}}.
}
