% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_creators.R
\name{format_authors}
\alias{format_authors}
\title{Format a list of author tables}
\usage{
format_authors(author_list, format_authors_function, ...)
}
\description{
Given a list (or one) author tables, format the authors per
the provided formatting function. Formatting functions currently include
\itemize{
\item \code{\link[=format_authors_as_text]{format_authors_as_text()}}
\item \code{\link[=format_authors_as_flextable]{format_authors_as_flextable()}}
}
}
