% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pubs.R
\name{format_authors_as_chunk}
\alias{format_authors_as_chunk}
\title{Format a data frame of authors as flextable output}
\usage{
format_authors_as_chunk(x)
}
\arguments{
\item{x}{}
}
\description{
Format a data frame of authors as flextable output
}
