% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_creators.R
\name{format_authors_as_flextable_chunk}
\alias{format_authors_as_flextable_chunk}
\title{Format a data frame of authors as flextable chunk output}
\usage{
format_authors_as_flextable_chunk(
  x,
  formatting_function = format_author_as_flextable_chunk
)
}
\arguments{
\item{formatting_function}{A function suitable for generating flextable chunk
output for an author. Default is \code{\link[=format_author_as_flextable_chunks]{format_author_as_flextable_chunks()}}.}
}
\value{
A flextable chunk (structured data frame)
}
\description{
Format a data frame of authors as flextable chunk output
}
\seealso{
\itemize{
\item \code{\link[=format_creators_as_flextable_chunk]{format_creators_as_flextable_chunk()}} for generic formatting function for creators.
}
}
