% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_creators.R
\name{format_authors_as_text}
\alias{format_authors_as_text}
\title{Format authors for text}
\usage{
format_authors_as_text(.x, collapse = ", ")
}
\arguments{
\item{.x}{A table of authors}

\item{collapse}{Character string between authors. Default is ', '}
}
\description{
Given an author table, format the output to include the list name
and values as a text-worthy chunk.
}
