% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_creators.R
\name{format_creators}
\alias{format_creators}
\title{Format creators of various types}
\description{
Creators exist across different reporting types, such as
grants (investigators), publications (authors) and
presentations (presenters, co-authors). These are all considered as
creators and much of the code associated with formatting the list of these
creators is the same.
}
\details{
Creators can be investigators, authors or presenters depending on the research
product. In any of these cases, however, the formatting is similar if not identical.

First, it is important to remember that a grant, publication or presentation table
consist of lists of creators. That is grant "A" has an investigator table unique
to this grant. Grant "B" also has a unique investigator table. Thus, when calling
\code{\link[=format_authors]{format_authors()}}, \code{\link[=format_presenters]{format_presenters()}} or \code{\link[=format_investigators]{format_investigators()}}, we should
expect a list of tables representing different research products.

From a list of tables, a formatting function is used to format the creators within
a research product as a single entity. Currently, this could be text or a flextable
"chunk". A set of functions (\code{format_authors_as_...}, \code{format_presenters_as_...},
\code{format_investigators_as_...})
expect a table of creators which are then formatted individually
(\code{format_author_as_...}, \code{format_presenters_as_...}, \code{format_investigators_as_...}) and
combined.

The approach taken for formatting creators is generally as follows (considering
the creator type author):
\itemize{
\item \link{format_authors(authors_list, author_formatting_function)}
\item \link{format_authors_as_flextable?}
\item \code{\link[=format_authors_as_text]{format_authors_as_text()}}
\item \code{\link[=format_author_as_text]{format_author_as_text()}}
}
}
