% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_creators.R
\name{format_creators_as_flextable_chunk}
\alias{format_creators_as_flextable_chunk}
\title{Format a list of creators into flextable structure}
\usage{
format_creators_as_flextable_chunk(x, formatting_function)
}
\arguments{
\item{formatting_function}{}
}
\description{
Given a table of creators (with the table consisting
of multiple columns, including the name, title, role, etc), consolidate
this to flextable-formatted output.
}
\details{
Flextable output is tabular but can involve multiple lines
of text per row of a table (to accommodate formatting, etc). This
function will apply the formatting function to a creator table (all of the
creators for a single research product) and then combine the flextable
output into a single flextable "chunk" suitable for adding to a flextable.
}
