% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_creators.R
\name{format_creators_as_text}
\alias{format_creators_as_text}
\title{Derive creator name and roles as single text}
\usage{
format_creators_as_text(.x, collapse = ", ", ...)
}
\description{
A creator (investigator, author, presenter) can have a partnership role and a
specific role. This function formats the information for output.
}
\details{
A creator is a person associated with a research product (grant, publication
or presentation). There can be one or more creators, and they may have different roles on
each product. Additionally, the creator has a partnership role which
may change over time. Thus, each research product has a tibble associated with the
creators for that work and include descriptions of their name,
role in the product (coI, author) and partnership role.

This function produces a text summary of this information so that it could
be printed out succinctly. There are many different parameters that can be
tweaked (see \code{format_investigator_name} for many formatting options passed on
through this function).

A creators table has one or more creators using standardized field
names (not yet an object, though). Given this table, we convert the fields
into a single string (per creator) then consolidate the creators into a single
string (using collapse to separate the entries).

The end of result of this function is a single string consolidating the
creators in \code{.x}.
}
\seealso{
\code{\link[=format_investigator_name]{format_investigator_name()}} for flags to format the investigator name.
}
