% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_format_flextable.R
\name{format_field}
\alias{format_field}
\title{Apply a formatting function to a flextable field conditionally.}
\usage{
format_field(ft, row, column, cf, ff = flextable::as_b)
}
\arguments{
\item{ft}{\itemize{
\item Flextable to operate on
}}

\item{row}{\itemize{
\item Current row of table to work on
}}

\item{column}{\itemize{
\item Name of column to modify
}}

\item{cf}{\itemize{
\item conditional function (only apply formatting if function is true)
}}

\item{ff}{\itemize{
\item formatting function (defaults to as_b)
}}
}
\value{
A modified flextable.
}
\description{
Apply a formatting function to a flextable field conditionally.
}
