% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{format_funding_mechanism}
\alias{format_funding_mechanism}
\title{Format the funding agency, program and mechanism}
\usage{
format_funding_mechanism(d)
}
\arguments{
\item{d}{A tibble representing the grants.}
}
\value{
A modified tibble with funding mechanism formatted.
}
\description{
Format the funding agency, program and mechanism
}
