% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_format_flextable.R
\name{format_grants_as_flextable_3}
\alias{format_grants_as_flextable_3}
\title{Title}
\usage{
format_grants_as_flextable_3(
  grants,
  show_grant_status = TRUE,
  show_u54_year = TRUE
)
}
\arguments{
\item{investigators}{}
}
\value{

}
\description{
Title
}
