% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_format_flextable.R
\name{format_grants_flextable}
\alias{format_grants_flextable}
\title{Format grants and investigators for MS-Word reporting.}
\usage{
format_grants_flextable(grants, investigators)
}
\arguments{
\item{d}{A composite (grant, investigator) object to report on.}
}
\value{

}
\description{
This function does multiple things with respect to formatting the output
of a grant table. It is assumed that the grant table has been filtered if
needed. This table will then be annotated with investigator information and
formatted using flextable.
}
