% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_text.R
\name{format_grants_text}
\alias{format_grants_text}
\title{Format grants/investigators to a text-printable output.}
\usage{
format_grants_text(grants)
}
\arguments{
\item{grants}{The grants table.}

\item{investigators}{The investigators table.}
}
\value{
A formatted table with a subset of columns for printing.
}
\description{
Format grants/investigators to a text-printable output.
}
\details{
The grants and investigators consist of many fields that
can be compressed and rearranged for nice output. This function
attempts to format the grant/investigators in a way that is
suitable for text-based output (e.g., a spreadsheet).
}
