% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_investigators.R
\name{format_investigator_group_as_flextable}
\alias{format_investigator_group_as_flextable}
\title{Format an investigator group as a flextable chunk}
\usage{
format_investigator_group_as_flextable(
  investigators,
  label = "",
  label_sep = "\\n",
  sep = "\\n",
  summary_field = "Investigator Summary",
  ...
)
}
\arguments{
\item{label}{A text string or a tibble containing one value}

\item{label_sep}{Separator between label and investigators (default: newline)}

\item{sep}{Separator between investigators (default: newline)}

\item{i}{An investigator tibble}
}
\value{
A flextable chunk representing the group.
}
\description{
An investigator group is a tibble of investigators that are considered a
grouping, for the purposes of reporting. Examples of groups include an
institution, or by Role. The goal of this function is to format this
group of investigators into a list of flextable "chunks".
}
\details{
The investigator group (a tibble of related investigators) should be
formatted as a list of flextable chunks. A group label can be applied
to these investigators, potentially formatted itself.

The \code{label} field is an overloaded field. It is either a simple text string
representing the name of the group, or a tibble. If it is a tibble, it is
expected to contain a single value representing the label to use. The parameter
is constructed this way to enable the \code{\link[dplyr:group_map]{dplyr::group_map()}} function to work,
since it passes a tibble with the grouped variables.
}
