% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_investigators.R
\name{format_investigator_group_as_text}
\alias{format_investigator_group_as_text}
\title{Format investigator groups as text chunks.}
\usage{
format_investigator_group_as_text(
  investigators,
  label = "",
  label_sep = ": ",
  sep = "; ",
  summary_field = "Investigator Summary",
  ...
)
}
\arguments{
\item{i}{The investigator group}
}
\value{

}
\description{
Given an investigator group, format the output to include the list name
and values as a text chunk.
}
\details{
The format of the input should be \code{group}=\code{string}. This creates a text summary
by combining the \code{group} and \code{string} into a single string (separated by newline).
}
