% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_investigators.R
\name{format_investigators}
\alias{format_investigators}
\title{Title}
\usage{
format_investigators(
  investigators_list,
  format_investigator_group,
  combine_investigator_groups,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
