% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_presentations.R
\name{format_presentation_as_paragraph}
\alias{format_presentation_as_paragraph}
\title{Format a presentation in flextable style}
\usage{
format_presentation_as_paragraph(...)
}
\arguments{
\item{x}{A list with the necessary named elements.}
}
\description{
Formats a single presentation as a flextable paragraph.
}
\details{
This function will create a \code{\link[flextable:as_paragraph]{flextable::as_paragraph()}} structure for
use in a flextable. It achieves this by creating a list of individual elements
and combining them together. These elements can be complex, formatted, etc.
}
\examples{
# NB format_presentation_as_flextable_paragraph
}
