% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_creators.R
\name{format_presenter_as_flextable_chunk}
\alias{format_presenter_as_flextable_chunk}
\title{Format a single presenter as a list of flextable chunks}
\usage{
format_presenter_as_flextable_chunk(...)
}
\arguments{
\item{...}{A list of variables corresponding to the presenter (name, affiliation, etc).}
}
\value{
A data frame of
}
\description{
Format a presenter as a table of
formatted chunks suitable for inserting into a \code{\link[flextable:as_paragraph]{flextable::as_paragraph()}}.
}
