% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_creators.R
\name{format_presenters}
\alias{format_presenters}
\title{Format a list of presenter tables}
\usage{
format_presenters(presenter_list, format_presenters_function, ...)
}
\arguments{
\item{presenter_list}{A list of presenter tables to format}

\item{format_presenters_function}{Function to format a presenters table into
a single formatted unit (text or otherwise).}

\item{...}{Other arguments to pass to \code{\link[pgimportr:format_name]{pgimportr::format_name()}}}
}
\value{
A list (same length as presenter_list) of formatted strings representing the
presenters.
}
\description{
Given a list (or one) presenter tables, format the presenters per
the provided formatting function. Formatting functions currently include
\itemize{
\item \code{\link[=format_presenters_as_text]{format_presenters_as_text()}}
\item \code{\link[=format_presenters_as_flextable]{format_presenters_as_flextable()}}
}
}
