% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_publications.R
\name{format_publications_as_flextable}
\alias{format_publications_as_flextable}
\title{Format publications as flextable output}
\usage{
format_publications_as_flextable(x)
}
\arguments{
\item{x}{A data frame consisting of publication fields (and possibly other things).}
}
\value{
A list of formatted publications.
}
\description{
Parses a data frame of publication information into a
formatted publication suitable for displaying in a \code{\link[flextable:flextable]{flextable::flextable()}}.
}
\details{
This function is actually just a wrapper for the formatting routine,
see \code{\link[=format_publication_as_flextable]{format_publication_as_flextable()}} for details. However, there is a bit
of boxing/unboxing to deal with, which this function manages. Use this
function when you have a data frame full of publications to format.
}
