% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_format_flextable.R
\name{format_table_conditionally}
\alias{format_table_conditionally}
\title{Format a flextable conditionally.}
\usage{
format_table_conditionally(
  ft,
  column,
  cf = function() {     TRUE },
  ff = flextable::as_b
)
}
\arguments{
\item{ft}{\itemize{
\item Flextable to operate on
}}

\item{column}{\itemize{
\item Name of column to modify
}}

\item{cf}{\itemize{
\item conditional function (only apply formatting if function is true)
}}

\item{ff}{\itemize{
\item formatting function (defaults to as_b)
}}
}
\value{

}
\description{
Format a flextable conditionally.
}
