% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{get_docx_template}
\alias{get_docx_template}
\title{Load docx template}
\usage{
get_docx_template(tag = "monthly")
}
\arguments{
\item{tag}{The template (currently monthly or yearly).}
}
\value{
A docx object based on the template.
}
\description{
For reporting, there are several templates available for generating
reports. This function will load the template (selected by the tag name)
and return it for use by officer, etc.
}
\details{
The template is particularly important because it defines various styles that
are assumed to exist and be configured a certain way for the needed look. These
template styles can be altered in the template documents, thereby changing the
look of the report relatively straightforwardly.
}
