% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_field_choices}
\alias{get_field_choices}
\title{Return the choices for a field}
\usage{
get_field_choices(dictionary, field)
}
\arguments{
\item{field}{}
}
\description{
Return the choices for a field
}
