% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_publications.R
\name{import_publications}
\alias{import_publications}
\title{Import publications from REDCap database}
\usage{
import_publications(uri, token)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
Import publications from REDCap database
}
