% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{import_redcap_data}
\alias{import_redcap_data}
\title{Import REDCap export data into R for reporting.}
\usage{
import_redcap_data(f)
}
\arguments{
\item{f}{The filename for importing}
}
\value{
A tibble with grant data available.
}
\description{
The REDCap data can be exported in R format. This consists of an
R script and corresponding CSV to read in. Labels are defined,
factors are created, etc.
}
\details{
There are a few caveats to using this type of data which this function
overcomes. The result is a similar data structure, but cleaned up for use
in this library.
}
