% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{info_box}
\alias{info_box}
\title{Create an info box}
\usage{
info_box(title, content, fill = "blue", col = "white")
}
\arguments{
\item{title}{A title (centered in the middle)}

\item{content}{Content to show (centered, larger than title)}

\item{fill}{Color for the background}

\item{col}{Color of the text}
}
\value{
A ggplot object that represents the info box.
}
\description{
Create an info box that is a color background with a title/content centered
in the middle.
}
\details{
This function implements the drawing of a "card", suitable for a dashboard. The idea
is that one creates a color background, then plots a title and text centered in the
plot window. The plot fills the entire boundary.
}
\examples{
\dontrun{
info_box("Title","Content", fill ="blue", col="white")
}
}
