% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_core_support}
\alias{is_core_support}
\alias{is_grant_core_support}
\alias{is_pub_core_support}
\title{Title}
\usage{
is_core_support(.x)

is_grant_core_support(.x)

is_pub_core_support(.x)
}
\arguments{
\item{.x}{}
}
\description{
Title
}
\section{Functions}{
\itemize{
\item \code{is_grant_core_support()}: Is grant supported by core

\item \code{is_pub_core_support()}: Is pub supported by core

}}
