% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_creator_esi}
\alias{is_creator_esi}
\alias{is_author_esi}
\alias{is_investigator_esi}
\title{Title}
\usage{
is_creator_esi(.x)

is_author_esi(.x)

is_investigator_esi(.x)
}
\arguments{
\item{.x}{}
}
\value{

}
\description{
Title
}
\section{Functions}{
\itemize{
\item \code{is_author_esi}: Is the author an ESI?

\item \code{is_investigator_esi}: Is the investigator an ESI?
}}

