% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_esi_investigator}
\alias{is_esi_investigator}
\title{Is investigator ESI?}
\usage{
is_esi_investigator(role)
}
\arguments{
\item{role}{The Role of the investigator tibble}
}
\value{
A logical indicating if the investigator is an ESI.
}
\description{
The Early Stage Investigator is a particular category of investigator to report
on. This returns the flag if this investigator is an ESI.
}
