% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_esi_related}
\alias{is_esi_related}
\alias{is_grant_esi_related}
\alias{is_pub_esi_related}
\title{Is the grant an ESI-related grant?}
\usage{
is_esi_related(.x = NULL)

is_grant_esi_related(...)

is_pub_esi_related(...)
}
\arguments{
\item{.x}{A grants table or nothing (in which case the field name is returned).}
}
\value{
A vector of logical values indicating if the grant is ESI-related or the field name.
}
\description{
This is an indicator if the grant is an ESI-related grant (a criteria
for reporting). In particular, an ESI-related grant is one in which any of the
key personnel listed in the database is an ESI.
}
\details{
Return a vector of logicals indicating if the grants represented by that
field are ESI-related or not (or the field name if tb is not included).
}
\section{Functions}{
\itemize{
\item \code{is_grant_esi_related}: Is grant related to an ESI?

\item \code{is_pub_esi_related}: Is publication related to an ESI?
}}

