% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_grant_funded}
\alias{is_grant_funded}
\title{Is the grant funded?}
\usage{
is_grant_funded(tb = NULL)
}
\arguments{
\item{f}{String (or vector of strings) with Grant Status}
}
\value{
Logical indicating if the grant has been funded.
}
\description{
Return a logical value for each grant status in the supplied list,
indicating if the grant is funded or not.
}
\details{
The grant table field to use is Grant Status
}
