% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_grant_in_preparation}
\alias{is_grant_in_preparation}
\title{Is the grant in preparation?}
\usage{
is_grant_in_preparation(tb = NULL)
}
\arguments{
\item{f}{A vector of logical values extracted from the is_grant_in_preparation field.}
}
\value{
A vector of logical values indicating if the grant is In Preparation.
}
\description{
The field is_grant_in_preparation indicates if the grant is in preparation.
Return a vector of logicals indicating if the grants represented by that
field are in preparation or not.
}
\details{
Note this function primarily exists as a method for documenting the use of the field
for users of the
library, not necessarily library developers.
}
