% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_grant_joint}
\alias{is_grant_joint}
\title{Is the grant a joint grant between the partner institutions?}
\usage{
is_grant_joint(tb = NULL)
}
\arguments{
\item{tb}{A grants table or nothing (in which case the field name is returned).}
}
\value{
A vector of logical values indicating if the grant is PHSU Cancer Related or field name.
}
\description{
This is an indicator if the grant is a joint grant between PHSU and Moffitt (a criteria
for reporting). Return a vector of logicals indicating if the grants represented by that
field are joint or not (or the field name if tb is not included).
}
