% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_grant_not_funded}
\alias{is_grant_not_funded}
\title{Is the grant not funded (a decision has been made)?}
\usage{
is_grant_not_funded(tb = NULL)
}
\arguments{
\item{f}{String (or vector of strings) with Grant Status}
}
\value{
Logical indicating if the grant has not been funded.
}
\description{
Return a logical value for each grant status in the supplied list,
indicating if the grant is officially not funded.
}
\details{
The grant table field to use is Grant Status

Note this differs from pending
review, in which the decision has not yet been made.
}
