% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_grant_pending_review}
\alias{is_grant_pending_review}
\title{Is the grant pending review?}
\usage{
is_grant_pending_review(tb = NULL)
}
\arguments{
\item{f}{String (or vector of strings) with Grant Status}
}
\value{
Logical indicating if the grant is pending review.
}
\description{
Return a logical value for each grant status in the supplied list,
indicating if the grant is pending review or not.
}
\details{
The grant table field to use is Grant Status
}
