% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_grant_phsu_cancer_related}
\alias{is_grant_phsu_cancer_related}
\title{Is the grant PHSU Cancer Related?}
\usage{
is_grant_phsu_cancer_related(tb = NULL)
}
\arguments{
\item{tb}{A grants table or nothing (in which case the field name is returned).}
}
\value{
A vector of logical values indicating if the grant is PHSU Cancer Related or field name.
}
\description{
This is an indicator if the grant is related to PHSU and Cancer (a criteria
for reporting). Return a vector of logicals indicating if the grants represented by that
field are PHSU Cancer-related or not (or the field name if tb is not included).
}
