% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_grant_rtype}
\alias{is_grant_rtype}
\title{Is the grant R type?}
\usage{
is_grant_rtype(tb = NULL)
}
\arguments{
\item{tb}{A grants table or nothing (in which case the field name is returned).}
}
\value{
A vector of logical values indicating if the grant is Rtype or field name.
}
\description{
This is an indicator if the grant is an Rtype grant (e.g., R01, R03)
Return a vector of logicals indicating if the grants represented by that
field are R-type or not (or the field name if tb is not included).
}
