% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_grant_submitted}
\alias{is_grant_submitted}
\title{Is the grant submitted?}
\usage{
is_grant_submitted(tb = NULL)
}
\arguments{
\item{f}{String (or vector of strings) with Grant Status}
}
\value{
Logical indicating if the grant is submitted.
}
\description{
Return a logical value for each grant status in the supplied list,
indicating if the grant is submitted or not.
}
\details{
The grant table field to use is Grant Status
}
