% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_new_investigator}
\alias{is_new_investigator}
\title{Is the role a New Investigator role?}
\usage{
is_new_investigator(x)
}
\arguments{
\item{x}{The data to check for being new investigator(s)}
}
\value{
Logical value indicating if role is a new investigator.
}
\description{
Note this is vector-aware, so either a single value or a vector
of investigators can be used.
}
