% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_other_support}
\alias{is_other_support}
\alias{is_grant_other_support}
\alias{is_pub_other_support}
\title{Title}
\usage{
is_other_support(.x)

is_grant_other_support(.x)

is_pub_other_support(.x)
}
\arguments{
\item{.x}{}
}
\description{
Title
}
\section{Functions}{
\itemize{
\item \code{is_grant_other_support()}: Is grant supported by other support

\item \code{is_pub_other_support()}: Is pub supported by other support

}}
