% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u54.R
\name{is_phsu_cancer_related}
\alias{is_phsu_cancer_related}
\alias{is_pub_phsu_cancer_related}
\alias{is_grant_phsu_cancer_related}
\title{Is the pub PHSU Cancer Related?}
\usage{
is_phsu_cancer_related(.x)

is_pub_phsu_cancer_related(.x)

is_grant_phsu_cancer_related(.x)
}
\arguments{
\item{.x}{A publication table or nothing (in which case the field name is returned).}
}
\value{
A vector of logical values indicating if the pub is PHSU Cancer Related or field name.
}
\description{
This is an indicator if the pub is related to PHSU and Cancer (a criteria
for reporting). Return a vector of logicals indicating if the grants represented by that
field are PHSU Cancer-related or not (or the field name if .x  is not included).
}
\section{Functions}{
\itemize{
\item \code{is_pub_phsu_cancer_related}: Is publication PHSU Cancer-related

\item \code{is_grant_phsu_cancer_related}: Is grant PHSU Cancer-related
}}

