% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_role}
\alias{is_role}
\alias{is_current_esi_investigator}
\alias{is_new_investigator}
\alias{is_trainee}
\title{Does investigator have a target role}
\usage{
is_role(role, target)

is_current_esi_investigator(role)

is_new_investigator(role)

is_trainee(role)
}
\arguments{
\item{target}{}
}
\value{
A logical indicating if the investigator is the role
}
\description{
Does investigator have a target role
}
\section{Functions}{
\itemize{
\item \code{is_current_esi_investigator()}: Is investigator ESI?

\item \code{is_new_investigator()}: Is the role a New Investigator role?

\item \code{is_trainee()}: Is the role a trainee?

}}
