% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_supported_by}
\alias{is_supported_by}
\alias{is_grant_supported_by}
\alias{is_pub_supported_by}
\alias{is_presentation_supported_by}
\title{Title}
\usage{
is_supported_by(.x = NULL, support)

is_grant_supported_by(.x, ...)

is_pub_supported_by(.x, ...)

is_presentation_supported_by(.x, ...)
}
\arguments{
\item{support}{}
}
\description{
Title
}
\section{Functions}{
\itemize{
\item \code{is_grant_supported_by()}: Is grant supported by a resource?

\item \code{is_pub_supported_by()}: Is publication supported by a resource?

\item \code{is_presentation_supported_by()}: Is presentation supported by a resource?

}}
