% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_trainee}
\alias{is_trainee}
\title{Is the role a trainee?}
\usage{
is_trainee(x)
}
\arguments{
\item{x}{The data to check for being a trainee.}
}
\value{
Logical value indicating if role is a trainee.
}
\description{
Note this is vector-aware, so either a single value or a vector of
investigator roles can be used.
}
