% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_with_datestamp.R
\name{knit_docx_with_datestamp}
\alias{knit_docx_with_datestamp}
\title{Title}
\usage{
knit_docx_with_datestamp(input, output_dir = "delivery", ...)
}
\arguments{
\item{input}{Input file}

\item{output_dir}{Output directory (default: delivery)}

\item{...}{}
}
\value{

}
\description{
Title
}
\details{
This code was taken from
https://bookdown.org/yihui/rmarkdown-cookbook/custom-knit.html as a way
to change the output file associated with markdown from within R Studio.
}
