% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{map_checkbox}
\alias{map_checkbox}
\title{Map a checkbox variable into levels}
\usage{
map_checkbox(.x, dictionary, target, prefix = "manu_supported_cores")
}
\arguments{
\item{prefix}{}
}
\value{

}
\description{
Map a checkbox variable into levels
}
