% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{map_checkbox_choice}
\alias{map_checkbox_choice}
\title{Map checkbox choice to label}
\usage{
map_checkbox_choice(.x, .choices)
}
\arguments{
\item{.choices}{}
}
\value{

}
\description{
A checkbox choice is a field that can be either Checked
or Unchecked. Transform this entry to the corresponding label text.
}
\details{
In REDCap, a checkbox is represented as a series of fields
with the choice index appended. That is, it is of the form\preformatted{checkbox___1
}

The data dictionary (metadata) provides a lookup of what 1 is labeled. The
content of the field is NA, Checked or Unchecked.

This function will take a column of data from a tibble assumed to be in
checkbox form. It requires a tibble of choices, to map the Checked
value to a corresponding Label value.
}
