% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_noNA}
\alias{paste_noNA}
\title{Paste two vectors together using sep, suppressing NA (rather than contaminating).}
\usage{
paste_noNA(v1, v2, sep = ", ")
}
\arguments{
\item{v1}{String (or vector of strings)}

\item{v2}{String (or vector of strings)}

\item{sep}{Separator (e.g., ',')}
}
\value{
A string (or vector of strings) with v1 and v2 pasted together, avoiding NA's.
}
\description{
Paste two vectors together using sep, suppressing NA (rather than contaminating).
}
\details{
Paste and str_c both appear to have a problem combining two strings when one is NA.
Or alternately, when one string is NA the result should be only the first name with
no sep or second component. This function does that correctly in the case of NA.
}
