% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partnership_grant.R
\name{pg_last_year}
\alias{pg_last_year}
\title{Return the last year of the Partnership Grant}
\usage{
pg_last_year()
}
\value{
An integer represent the last grant year (e.g. 10).
}
\description{
Return the last year of the Partnership Grant
}
