% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u54.R
\name{pg_start_date}
\alias{pg_start_date}
\title{Given a Partnership Grant funding year, return the start date of that year.}
\usage{
pg_start_date(yr = "Y01")
}
\arguments{
\item{yr}{The funding year - default is the initial grant year ("Y01")}
}
\value{
The start date (as a lubridate) of the funding year.
}
\description{
Given a Partnership Grant funding year, return the start date of that year.
}
