% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partnership_grant.R
\name{print_pg_years}
\alias{print_pg_years}
\title{Return the text version of Parntership Grant years (e.g., Y01) for all years.}
\usage{
print_pg_years()
}
\value{
A list of grant years in order.
}
\description{
Return the text version of Parntership Grant years (e.g., Y01) for all years.
}
