% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{read_redcap_data}
\alias{read_redcap_data}
\title{Read RedCAP data into memory.}
\usage{
read_redcap_data(f)
}
\arguments{
\item{f}{The filename to read from.}
}
\value{
A tibble representing the redcap data.
}
\description{
Given that this data is in R format, it is actually a script
to interpret. However, we do a few things to that script to
make it easier to process.
}
