% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_redcap_data_online}
\alias{read_redcap_data_online}
\title{Retrieve REDCap data from the server}
\usage{
read_redcap_data_online(uri, token)
}
\arguments{
\item{token}{}
}
\description{
Using the REDCapR library, import the raw and labeled versions of
a project. Also download the data dictionary.
}
