% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{remove_investigators_table}
\alias{remove_investigators_table}
\title{Remove investigator entries from a tibble.}
\usage{
remove_investigators_table(d)
}
\arguments{
\item{d}{A tibble containing investigators and grants.}
}
\value{
A modified tibble with investigator entries removed.
}
\description{
Remove investigator entries from a tibble.
}
