% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{rename_field_from_checkbox_index}
\alias{rename_field_from_checkbox_index}
\title{Title}
\usage{
rename_field_from_checkbox_index(x, dictionary)
}
\arguments{
\item{dictionary}{}
}
\description{
Title
}
