% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_joint_grants_as_flextable}
\alias{report_joint_grants_as_flextable}
\title{Filter and format a grant table}
\usage{
report_joint_grants_as_flextable(grants, ...)
}
\arguments{
\item{d}{A grant table}
}
\description{
Given a tibble associated with selected grants, format the tibble as a
flextable object to insert into an Office document. This particular function
manages the "joint grant" style.
}
\details{
The model for this package is to provide a fairly comprehensive tibble with
grant information. There are filters to subset to a given list of grants. Once
this is done, we want to generate pretty output representing the table. This
function will provide a nicely-formatted flextable suitable for reporting
requirements.

Of note, this function will apply filters for the specific condition (e.g., joint
grants) as implemented by the \verb{filter_*} approach in the library.
}
