% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.r
\docType{data}
\name{role_order}
\alias{role_order}
\title{The native order of investigator roles within the system.}
\format{
An object of class \code{character} of length 20.
}
\usage{
role_order
}
\description{
The native order of investigator roles within the system.
}
\details{
Investigator roles follow a hierarchy from PI through consultant and mentor. This variable consists
of all known roles in order so that investigators can be arranged using this hierarchy.
}
\keyword{datasets}
