% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{split_choice_levels}
\alias{split_choice_levels}
\title{Split choice levels in discrete columns}
\usage{
split_choice_levels(s)
}
\arguments{
\item{s}{A string to split into Value and Label (vectorized).}
}
\value{
A tibble with \code{Value} and \code{Label} fields.
}
\description{
REDCap choices are in the form of N, VALUE. This separates out
the two components, trims them and returns them as a tibble.
}
