% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_date_range}
\alias{split_date_range}
\title{Split dates that are in the form of start_date - end_date.}
\usage{
split_date_range(s)
}
\arguments{
\item{s}{A character vector representing the combined date.}
}
\value{
Two character vectors representing the first and second date, respectively.
}
\description{
Split dates that are in the form of start_date - end_date.
}
\examples{
\dontrun{
split_date_range("2024_01_01 - 2024_01_02")
     [,1]          [,2]
[1,] "2024_01_01 " " 2024_01_02"
}
}
