% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_format_flextable.R
\name{str_includes_tag}
\alias{str_includes_tag}
\title{Does the given string include a special tag (^n^)}
\usage{
str_includes_tag(str)
}
\arguments{
\item{str}{}
}
\value{

}
\description{
To allow for formatting of strings in a table format, we insert
special tags (^n^) that can be found and replaced with formatting
once we layout the table.
}
