% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_format_flextable.R
\name{str_replace_tag_superscript}
\alias{str_replace_tag_superscript}
\title{Replace a special tag with superscript formatting.}
\usage{
str_replace_tag_superscript(str)
}
\arguments{
\item{str}{}
}
\value{

}
\description{
Replace a special tag with superscript formatting.
}
