% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_grants_as_flextable_delta}
\alias{style_grants_as_flextable_delta}
\title{Title}
\usage{
style_grants_as_flextable_delta(d, ...)
}
\arguments{
\item{d}{}
}
\value{

}
\description{
Title
}
\details{
This function does no filtering to the table, but does the work of styling the
table as a flextable. This can involve some heavy work, however, given the various
ways in which styling requires post-processing.
}
