% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_grants_as_text_alpha}
\alias{style_grants_as_text_alpha}
\title{Format grants/investigators to a text-printable output.}
\usage{
style_grants_as_text_alpha(grants, ...)
}
\arguments{
\item{grants}{The grants table.}

\item{...}{Other parameters to pass to format_investigators.}
}
\value{
A formatted table with a subset of columns for printing.
}
\description{
Format grants/investigators to a text-printable output.
}
\details{
The grants consist of many fields that
can be compressed and rearranged for nice output. This function
attempts to format the grant/investigators in a way that is
suitable for text-based output (e.g., a spreadsheet).
}
