% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_presentations.R
\name{style_presentations_as_flextable_delta}
\alias{style_presentations_as_flextable_delta}
\title{A styling function for presentations}
\usage{
style_presentations_as_flextable_delta(d, ...)
}
\arguments{
\item{d}{A data frame of presentations (typed as presentations)}
}
\value{
A flextable consisting of formatted publication entries.
}
\description{
A styling function for presentations
}
\details{
This function does no filtering to the table, but does the work of styling the
table as a flextable. This can involve some heavy work, however, given the various
ways in which styling requires post-processing.

The flextable (and text) styles are listed by alpha, delta, to distinguish them
as we are determining the best formatting approach.
}
\seealso{
\itemize{
\item \link{style} for general styling information.
\item \link{style_presentations} for general presentation style information.
\item \link{style_publictions} for publication styling.
\item \link{style_grants} for grant styling.
}
}
