% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_publications.R
\name{style_publications_as_text_alpha}
\alias{style_publications_as_text_alpha}
\title{Format publications/authors as text-printable output}
\usage{
style_publications_as_text_alpha(.x, ...)
}
\arguments{
\item{.x}{The publications table}

\item{...}{Other parameters to pass to format_investigators.}
}
\value{
A formatted table with a subset of columns for printing.
}
\description{
Format publications/authors as text-printable output
}
\details{
The publications consist of many fields that
can be compressed and rearranged for nice output. This function
attempts to format the publications/authors in a way that is
suitable for text-based output (e.g., a spreadsheet).
}
