% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_pubs_as_flextable_delta}
\alias{style_pubs_as_flextable_delta}
\title{A styling function for publications}
\usage{
style_pubs_as_flextable_delta(d, ...)
}
\arguments{
\item{d}{A data frame of publications}
}
\value{
A flextable consisting of formatted publication entries.
}
\description{
A styling function for publications
}
\details{
This function does no filtering to the table, but does the work of styling the
table as a flextable. This can involve some heavy work, however, given the various
ways in which styling requires post-processing.

The flextable (and text) styles are listed by alpha, delta, to distinguish them
as we are determining the best formatting approach.
}
