% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_publications.R
\name{style_pubs_as_flextable_gamma}
\alias{style_pubs_as_flextable_gamma}
\title{Title}
\usage{
style_pubs_as_flextable_gamma(d, ...)
}
\arguments{
\item{d}{}
}
\description{
Title
}
\details{
This function does no filtering to the table, but does the work of styling the
table as a flextable. This can involve some heavy work, however, given the various
ways in which styling requires post-processing.
}
