% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_unite}
\alias{vec_unite}
\title{Unite parallel vectors of strings into a combined single vector}
\usage{
vec_unite(..., sep = ", ", na.rm = TRUE, na.empty = TRUE)
}
\arguments{
\item{...}{Vectors of strings to combine.}

\item{sep}{Used \code{sep} for combining elements.}

\item{na.rm}{Remove NA values before combining.}

\item{na.empty}{Treat empty ("") values as NA.}
}
\value{
A vector of combined strings.
}
\description{
Unite parallel vectors of strings into a combined single vector
}
