% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_core_supported}
\alias{filter_core_supported}
\alias{filter_grants_core_supported}
\alias{filter_pubs_core_supported}
\title{Filter table on those with any core support.}
\usage{
filter_core_supported(.x)

filter_grants_core_supported(.x)

filter_pubs_core_supported(.x)
}
\arguments{
\item{d}{The grant table}
}
\value{
A filtered grant table with only grants having grant support.
}
\description{
Filter table on those with any core support.
}
\section{Functions}{
\itemize{
\item \code{filter_grants_core_supported}: Filter grants with core support

\item \code{filter_pubs_core_supported}: Filter pubs with core support
}}

