% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u54.R
\name{pg_last_year}
\alias{pg_last_year}
\title{Return the last year of the Partnership Grant}
\usage{
pg_last_year()
}
\value{
An string representing the last (max) grant year.
}
\description{
Return the last year of the Partnership Grant
}
