% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{split_choices}
\alias{split_choices}
\title{Split multiple choices into discrete fields}
\usage{
split_choices(s)
}
\arguments{
\item{s}{A string to split into choices (vectorized).}
}
\value{
A tibble of \code{Value} and \code{Label} fields for all choices.
}
\description{
A selection is a series of choices that can be selected. This
function extracts them as a tibble (since each choice is itself coded).
}
